+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.2.2 [unreleased]
 
+- Added bubble chart for Chart.js
 - Limit concurrent requests to avoid overloading the server
 
 ## 2.2.1